requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
Django==4.2.6
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
typing-extensions==4.8.0
validators==0.22.0
